Added optimzed version of WeightedInner.
Optimized the routine WeightedInner by working directly on the
gsl_vector_complex data structures. This removed an unnecessary vector
copy, merged several loops into one (reducing loop overhead and memory
traffic) and allowed the compiler to vectorize (when compiling with
the correct compiler options).

I left the old routine in the file but renamed it to WeightedInnerOld.